refactor use module
