+++ lib/custom_log_space/log_writer.rb
@@ -1,10 +1,13 @@
 # frozen_string_literal: true
 
 module CustomLogSpace
-  # The LogWriter class provides methods for writing log messages to custom log files.
+  # The LogWriter module provides methods for writing log messages to custom log files.
   # It allows the creation of log directories, handling file errors, and appending messages to log files.
-  class LogWriter
-    def self.write_to_custom_log(directory_path, message)
+  module LogWriter
+    private
+
+    def write_to_custom_log(message)
+      directory_path = controller_log_directory
       FileUtils.mkdir_p(directory_path) unless Dir.exist?(directory_path)
       custom_log_path = custom_log_file_path(directory_path)
 
@@ -16,15 +19,26 @@ module CustomLogSpace
       handle_file_error(e)
     end
 
-    def self.custom_log_file_path(directory_path)
+    def handle_file_error(error)
+      error_prefix = error.is_a?(SystemCallError) ? "Error" : "IO Error"
+      puts "#{error_prefix}: #{error.message}"
+    end
+
+    def custom_log_file_path(directory_path)
       action_name = Thread.current[:current_action]
       log_file_name = "#{action_name}.log"
       File.join(directory_path, log_file_name)
     end
 
-    def self.handle_file_error(error)
-      error_prefix = error.is_a?(SystemCallError) ? "Error" : "IO Error"
-      puts "#{error_prefix}: #{error.message}"
+    def custom_log_directory
+      today = Time.now.strftime("%Y%m%d")
+      time = Time.now.strftime("%H%M")
+      File.join(Rails.root, "log", "custom_log_space", today, time)
+    end
+
+    def controller_log_directory
+      controller_name = Thread.current[:current_controller].underscore
+      File.join(custom_log_directory, controller_name)
     end
   end
 end
